requirements.txt
autopep8==1.6.0
pycodestyle==2.8.0
python-dateutil==2.8.2
six==1.16.0
toml==0.10.2
